Checks to see if the word is in a String
If the word subset in a string is Valid, it will be concatted to an
empty string, let’s call it filledString. After it is looped through,
the length of the submitted word is compared to the filledString. If
the string lengths are the same, all words are valid, if not, the
string lengths will be different and it will not work.